Change the way bars are added and removed
Refactored user and bar schemas so that bars now hold an array of
attendees, instead of using keeping an array of bars they're attending.

In addition, if the user is logged in, the bar search endpoint marks
the bars they're attending before sending the response. This is done by
adding a "userAttending" property to bars they're attending and setting
it to true.